python/pyproject.toml
[tool.poetry]
name = "idris-dl-python"
version = "0.1.0"
description = ""
authors = ["Benjamin Miller <benmiller179@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
numpy = "^2.1.3"
pandas = "^2.2.3"
pyarrow = "^18.0.0"
pillow = "^11.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
